docs: 📝 example station haldensleben
The example folder contains now the tracktopolgy of the station
Haldensleben. The data was extracted from OSM and contains all necessary
information to create a proper track model.

For the MVP only a subset is really necessary: TrackId, Length (if it
is missing it needs to be calculated). The connectivity is not encessary
for the MVP sind it is implicitely given later by the track list in the
route definitions.