docs: 📝 example station (track topology)
An example file containg the track topology of the station Haldensleben
is provided. The track topology was extracted from OSM and contains a
lot of information not necessary for the MVP.

The MVP will only need the length of a track. If it is missing it needs
to be computed during conversion. Also the tracklabels need to be kept
since the user will assign these labels the track type during the
configuration. In principle a graph representation of the track network
can be created. But for the MVP the connectivity is given by the
definition of the routes. Routes contain a list of tracks in the proper
order. Furher imnfomration and information how to reduce the data is
given in the provided README.