docs: 📝 initial architecture and dev docs. (#35)
* docs: 📝 initial architecture and dev docs.

This commit contains the initial architecture documentation according to
arc42 and c4. The documentation can be found in the docs/architecture
folder. The bounded-contextes in the corresponding subfolder were
identified in the event storming sessions within the skydeck
acceleator.

The documentation for the developers is added and can be
found in the docs/development folder.

Definition of the requierementes are in the docs/requirements foldern.
The contained use cases are based on the concept of Frederik Schäfer.

* docs: 📝 function for tracks and route definition

The user can specify the different type of tracks directly for the
workshop. Also sepcification of the routes a loco can drive is
defined

* docs: 📝 background info for config validation