feat(osm): 🎉 extract and visualize OSM data.
A new tool allows to extract, clip, project and plot OSM data and is
used for railway networks (strictly it is not limited to it).

For extraction the user specifies a bounding box or a polygon. In its
interior the specified OSM data is extracted. The Overpass API is used
with a recursive query to follow e.g. railroad tracks starting in the
interior, but ending outside the extraction area. Also this will take
intermediate data nodes into account. Therefore, data might
extend beyond the specified extraction area.

Clipping functionality is provided to reduce the the amount of data extending
from the extraction aread. It is tested which nodes and edges lie in the
extraction area. If an edge starts in the interior, travels outside and ends in
the interior again it will be kept. Also if the edge starts in the interior and
ends in the exterior it is still kept.

When trying to visualize the data it might be disorted due to the missing
projection. It is possible to apply elliptic Mercator projection to the data
which will add X and Y coordinates to the data.

For visualization a plotting capability based on MatplotLib is introduced.
Currently it assumes that it is used for visualization of railway data.
Switches and buffer stop are separately marked by a symbol.

Control of the functionality is done using a CLI. Further information is
available in the provided README.md

Unit tests and integration tests are provided leading to an overall coverage
roughly greater than 80%.

The existing project configuration (pyproject.toml) is extended to treat the
OSM extractor as a separate package with a cli. During the test with PyLint it
complied about the identation even ruff had formatted the files. Therefore,
configuration was adopted to work consistently to ruff.