chore(backend): 🐛 fix pre-commit for Windows.
The existing setup script for installing the precommit-hooks did not
work on Windows due to issues in character encoding and issues in the
regex. The file is adopted to our coding style.

The configuration of the mixed-line ending of the pre-commit hook is
altered. Detection / setting of the line ending is removed since it is
driven by the .gitattributes file.

This fixes #111