chore(backend): 🔧 improve cfg of backend
The configuration of MyPy is moved from the mypy.ini to the
pyproject.toml to have a centralized configuration. Support for pydantic using the corresponding plugin is added. Files in the test folder are now ignored.

The prior configuration defines two dev environments. One in project.optional-dependencies and one in the dependency-groups section. The latter is recommended for proper Python packaging. Therefore, all dev dependencies in the project.optional-dependencies are moved to the dependency-groups. Now all dev dependencies can be installed by using the --dev flag. The --all-extras flag is not necessary anymore.

Pydantic support for Pylint is added.