fix: 🐛 fix linting errors
During fixing of the liniting errors it was found that Ruff and MyPy
collide for the decorators.py file. Ruff remove the call of settattr
and assigns the value directly to attribute, while MyPy prefers the use
of settattr. Therefore, the related rule of Ruff is deactivated for the
file.