fix(backend): 🐛 fix linting of simulation context
The linting errors of the simulation context are fixed. The constructor
of the Scenario is altered to make the routes mandatory. The types were
adopted also avoiding the possible runtime error later when trying to
access the length attribute. The attribute is only available when the
contructor gets a Routes object. If nothing was passed it would set an
empty list, which does not have the .length attribute. The missing
__init__.py of the simulation context is added since some import errors
just stem from the missing file.

The test of the simuation context experienced type errors due to
incompability of the former introduced FakeWagon and FakeRoutes. Instead
now MagicMock is used. Also the altering of module path is removed to
avoid undesired side-effects.