Start simpy environment adapter (#122)
* chore: some fixes for mypy

# Conflicts:
#	popupsim/backend/src/configuration/model_route.py
#	popupsim/backend/src/main.py

* chore: rebase to main

* refactor: rename some files

* chore: rebase to main

* chore: rebase to main

# Conflicts:
#	popupsim/backend/pyproject.toml
#	popupsim/backend/tests/unit/test_model_scenario.py
#	popupsim/backend/tests/unit/test_model_train.py
#	popupsim/backend/tests/unit/test_model_wagon.py
#	popupsim/backend/tests/unit/test_model_workshop.py

* chore: format as commented in review

* fix(backend): 🐛 fix linting of simulation context

The linting errors of the simulation context are fixed. The constructor
of the Scenario is altered to make the routes mandatory. The types were
adopted also avoiding the possible runtime error later when trying to
access the length attribute. The attribute is only available when the
contructor gets a Routes object. If nothing was passed it would set an
empty list, which does not have the .length attribute. The missing
__init__.py of the simulation context is added since some import errors
just stem from the missing file.

The test of the simuation context experienced type errors due to
incompability of the former introduced FakeWagon and FakeRoutes. Instead
now MagicMock is used. Also the altering of module path is removed to
avoid undesired side-effects.

* refactor: ♻️ use pyproject.toml from root

The developer can now call all tools form the project root instead of
stepping into the popupsim/backend folder. All files are adopted to fix
all MyPy issues.

Ruff, Pylint, MyPy and Pytest work properly also the precommit-hook and
pipeline is adopted to the new folder layout

* chore: addjust pyproject for coverage tool

* feat: use ScenarioConfig in Simulation

* feat: use simpy from main

---------

Co-authored-by: Jan-Hendrik Wendisch <jan-hendrik.wendisch@deutschebahn.com>