docs(pages): 📝 render documentation as website
Rendering of the markdown files is now done using mkdocs. For a pull
request a preview will be added to the request and automatically
removed if the PR was merged or closed. The structure of the docs folder
is changed since now the preparation of the overall architecture
documentation is added besides the one for the MVP. The reason for this
is that due to the reduction of scope directly in the documentation
issues arise due missing information which is needed to prepare the
code for future implementations after the MVP.