docs(arch): 📝 update documentation
The existing architecture documentation is widely updated. Especially
the contexts are renamed and the code snippets adopted to fit more
to the code. 

The documentation will now be rendered as a static website using 
mkdocs. The link to the page for the review will be shown as a 
deployment status in the pull request.

The existing READMe is adopted to fit more the current state of  the
repository