feat(topo): ✨OSM data to PopUpSim topology
PopUpSim needs topology of the railway network for e.g. shunting
operations and definition of yards. This data is normally provided by
dedicated systems within railway companies, but this data is generally
not freely available.

PopUpSim will use its own data format internally to represent tracks,
track sequences, switches and so on. Since currently no system providing
the necessary data is available, a tool is provided allowing to extract
data from openrailwaymap. The user specifies a bounding box or polygon
in which railway data is extracted. By using the CLI it is possible to
configure also if unused tracks are filtered out. Extraction of the data
fully relies on the overpass API. Data is returned as a JSON with nodes
and edges describing the topology. At nodes or edges tags maybe present
for additional information like if the node is a switch.

During extraction a recursive query is send to the overpass API. This is
necessary since nodes outside the extraction area are missing which lead
to missing edges (one node lies in the inside and one outside). The
railway network might extend the defined extraction area. Therefore, in
a next step a clipping is done.

The extracted and clipped coordinates are still given in lat/lon. A
projection using the full elliptical mercator projection is applied for
visualization.

A plotting tool (called plotter) is shipped with the osm2extractor to
visualize the extracted data using matplotlib.

The possible commandline arguments fo the cli are described in the
provided README.md

PopUpSim uses an own format. A provided tool called osm2popupsim will
do the conversion. It reads the raw file and creates the reformatted
topology file (sill node edge represnetation). Additionally a tracks
file is created which virtually a sequence of edges to a track. A track can
have a name, type and has an id. The user can assign functionality like
parking, wretrofit and so on for each of these tracks.

Visualization of the PopUpSim data is provided using the via. It is possible to
visualize the nodes, edges and switches or show coloured the type of assigned
functionality to the tracks. Output is done using Folium to show the data as an
overlay to the OSM data.

The full arguments of the osm2popupsim are described in the sipped README.md