refactor(backend): ♻️ improve architecture
Many parts of the code were relatively strongly coupled, especially
cross-context coupling existed. For the configuration context dtos
are introduced to decouple the reading of the data from the
models/entities. It allows for a two layer validation. The first
directly on read in the configuration context (like field types) and
validation dedicated to the entity, aggregate or domain later on.
These are also preparations to allow later to integrate different input
types using an adapter pattern.

The architecture of the analytics context is improved. Especially the
entities, value_objects and aggregates are now available. Additionally
events are introduced.

Changes in both contexts still affected the workshop context. Therefore
the context needed adoption.

Most of the tests run. Only test relying on a proper scenario input are
now failing since validation is performed. Therefore, it is just a
matter of providing proper testcases as input.