refactor(backend): ♻️ improve architecture.
The analytics context is refactored and its domain is enriched allowing stronger cohesion and weeker coupling to the other contextes.

The former models used by the configuration context are now proper entities in the workshop operations context. The dtos in the context are now fully integrated and the old legacy code is removed.

This is a further step to increase maintainability and to prepare the usage of different input methods like csv or an api for the configuration.

Currently some tests will fail due to validation issues of the input. This will be fixed later. Therefore, these test can now be marked that they are allowed to fail.