refactor(backend): 🏗️ improve arch for cfg context
The architecture is changed for the configuration context. By using
ports and adapters input to PopUpSim can now be provided using either
csv or json. Additionally the process from reading the input, putting
the data into the dtos to fill the e.g. entities now use a pipeline
flow. Validation is now done in the contexts itself and only basic
validation when reading.