refactor(backend): 🎨 standardize attributes
Attributes like the ids always had prefixes like wagon_id, train_id
and so on. Leading to calls like wagon.wagon_id. The prefix for the ids
is now removed easying the code later on.

The same is done e.g. or the workshop context. Now
workshop_station_usage is now just usage.