feat(backend): ✨ Add 4-layer stacked validator
A four-layered validator is added with combined stacking of errors.
All warnings and errors will be gathered and returned to the user
instad of failing immediately with the first error.

The first layer checks the syntax like fied formats, types and
required fields. The second layer validates the business rules within
enties. The third layer validates cross references and the consistency
of data. The last, the fourth layer validates operational constraints
and the readiness of the simulation.

The implementation resides in the shared context since the validator is
used by all contexts.