fix(backend): 🐛 partial batching deadlock
In case of a partial batch (less wagons then retrofit stations) the
simulation got stuck. If the batch was not complete the locomotive was
never triggered to pick them up. Currently the most staightforward
solution is to use a simpy resource also for the retrofitted track to
allowing proper interaction of proceses.

ADRs are added showing different approaches to solve the problem.
Currently also the queues have dual use since they are used for tracking
and modeling the process.