feat(backend): ✨MVP feature set
The code now contains all MVP features and test of the behaviour of the
code are already conducted. Testing is done for different scenarios
without coupling with different number of wagons, retrofit stations and
batch sizes. For a signle wagon also the timinigs of coupling and
decoupling gives a valid timeline.

The orchestrator.py is currently a god object, large and complex. It
needs be refactored. The warnings due to its complexity and branching
are currently switched off.

Pylint will argue about code similarities. Unfortunately it seems to be
impossible to switch these error of locally. The similarity will be 
vanish in the future if the validation of the input is fully implemented.