refactor(backend): 🏗️ shunting operation context
The orchestrator of the workshop operations context is a big god object
doing a lot of different things. It exists due to the restriction to
three contexts. Especially the workshop context gets massively
cluttered.

To improve the situation now the shunting operations are moved to a
separate context. The shunting context has its own entities. The added
locomotive is already prepared to treat diffrent types of couplers.

The contect will be enriched in the future e.g. allowing to assign
workers to the loco, add shifts and so on for the loco.