refactor(backend): 🏗️ add yard operations context
The god object orchestrator.py of the workshop operations context
implements a lot of functionality regarding a yard like sorting wagons.

A more general approach is integrated in the yards operation context. It
can be defined if a yard has a hump, workshops and/or parking areas
allowing to flexible define a yard. In the future also the track
topology will be added.