refactor(backend): 🏗️ add popup wokshop  context
The new context is related to the retrofit process in the popup
workshop. It allows to disitnguish this process from any general
workshop context. The code poritons of this context in the orchestrator
are removed und are fully integrated in the new context. Additionally
the metrics are now computed in the context if they directly belong to
it. The analystics module therefore is adopted to use the metric.