fix(backend): 🐛 fix simpy validation error.
After refactoring to the async analytics the code throwed odd validation
errors deep down in Simpy. The analytics context is refactored and
corrected and all old non-async stuff is wiped out. This was done since
it was assumed that this change in the code was causing the trouble.
Also it greately simplified the implementation.

But the error resulted from the RetrofitResult class because two
implementations existed. On in the popup_workshop (pydantic model) and
in the retrofit_processor. The mismatch caused Simpy to throw an
ValidationError whe the processor tried to create a RetrofitResult
with a reason that does not exist in the Pydantic model.