feat: ✨ Version for OpenRailDays
This version is a massive change compared to the MVP. It fixes several
issues:
- Cannot handle more than one train arrival. It was never checked again
after the first train arrival that another train arrived. Currently it is
now polled all 3 minutes to check for a new train arrival. This will be
changed in the avoid the polling approach.
- Issues when creating the rakes and batches occured leading to stuck
simulations.
- The wagons were never transported to parking since if one only wagon
for two retrofit bays was present the code waited until the batch size
is two. This could never happen since wagons are only transported to the
workshops if they are empty
- Many position tracking issues are fixing
- Having multiple collection tracks and parking tracks lead to filling
errors.
- The parking tracks were overfilled since the capacity between
calculation of the track capacity was wrongly handled. The capacity
is computed in meters and based on the given lengts of the wagons it is
computed how many wagons should Simpy allow.
- A basic frontend using streamlit is added
- Functionality is improved at the expense of code quality