fix(backend): 🐛 fix rake generator warning.
When simulation ends before all processes of Simpy are ended Generator
warnings occured. The reason is the usage of the while True loops. The
generator do not cleanlz exit throwing warnings in the terminal. By
adding the additional try and except block this case is now catched
removing the generator warning.