fix(backend): wrong rejected wagons count
The number of rejected wagons in the metrics calculation is fixed. The
computation before was entirely wrong and just substracted the number
of retrofitted wagons from the arrived wagons and also used some max
computation as guard. This was an old implementation when not all
values were available. This is now removed and the rejected wagon event
is used to determine the count.

Also the computation of the workshop utilization is wiped out since it
does not belong here.