fix(backend): 🐛 workshop utilization
The workshop utilization was falsely computed. It took the track_WSX
into accounts whcih are the tracks a workshop resides on leading to
completely wrong utilization values. These tracks are filtered out.
The code has to be checked why they are treated here anyhow.

Also the bay utilization is fixed. Not it is possible to get the
utilization of each bay in a workshop properly.