feat(sim): ✨ add working examples
Two examples for the new code are added. One gives 10 trains in two days
showing the retrofitting process.

The second example is basically the same, but the trains are more
closely sheduled and the collection tracks are very short to force
rejection of wagons due to capacity limitations on the collection
tracks.